Move files & Peel CliMainWorker
